+++ src/time_split/_compat.py
@@ -1,7 +1,8 @@
 """Vendored functions."""
 
+from collections.abc import Callable
 from functools import wraps
-from typing import Callable, ParamSpec, TypeVar
+from typing import ParamSpec, TypeVar
 from warnings import warn
 
 from rics.misc import tname