+++ tests/integration/test_sklearn.py
@@ -17,7 +17,7 @@ def test_sklearn(caplog):
     assert splitter.get_n_splits(array) == 2
 
     pandas_split = split_pandas(index.to_series(), schedule="1d", log_progress=False)
-    for (train_index, test_index), pandas_fold in zip(splitter.split(array), pandas_split):
+    for (train_index, test_index), pandas_fold in zip(splitter.split(array), pandas_split, strict=True):
         assert (pandas_fold.data.index == array[train_index]).all()
         assert (pandas_fold.future_data.index == array[test_index]).all()